services/simulation-engine/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
redis==5.0.1
PyYAML==6.0.1
jsonpath-ng==1.6.1
python-jose==3.3.0
requests==2.31.0
alembic==1.13.1
minio==7.2.7
